admin_dashboard/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.0.2
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
celery==5.3.4
websockets==12.0
aiofiles==23.2.1
Pillow==10.1.0
qrcode==7.4.2
pyotp==2.9.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1